Save HZ as an unsigned long integer
Number of clock ticks per second is given as a long integer by sysconf()
function. It is also saved as an unsigned  long integer in sar's data
files (sa_hz).

+++ common.c
@@ -48,7 +48,7 @@
 char units[] = {'s', 'B', 'k', 'M', 'G', 'T', 'P', '?'};
 
 /* Number of ticks per second */
-unsigned int hz;
+unsigned long hz;
 /* Number of bit shifts to convert pages to kB */
 unsigned int kb_shift;
 
@@ -641,7 +641,7 @@ void get_HZ(void)
 		perror("sysconf");
 	}
 
-	hz = (unsigned int) ticks;
+	hz = (unsigned long) ticks;
 }
 
 /*
